+++ lightbulb/stringview.py
@@ -64,7 +64,7 @@ class StringView:
         buff = []
         while self.index < len(self.text):
             char = self.text[self.index]
-            if (char == " " or char == "\n") and self.expect_quote is None:
+            if char == " " and self.expect_quote is None:
                 self.index += 1
                 return "".join(buff)
             elif not self.expect_quote and char in _quotes:
@@ -91,28 +91,28 @@ class StringView:
         if buff:
             return "".join(buff)
 
-    def deconstruct_str(self) -> typing.List[str]:
+    def deconstruct_str(self, *, max_parse=float("inf")) -> typing.Tuple[typing.List[str], str]:
         """
         Method called to deconstruct the string passed in at instantiation into
         a list of all of its arguments. Arguments are defined as words separated by whitespace,
         or as words enclosed in quotation marks.
 
-        Raises:
-            :obj:`~.errors.UnclosedQuotes`: When an unexpected EOF is encountered
+        Keyword Args:
+            max_parse (:obj:`int`): The maximum number of arguments to parse. Defaults to ``float("inf")``.
 
+        Raises:
+            :obj:`~.errors.UnclosedQuotes`: When an unexpected EOF is encountered.
 
         Returns:
-            List[ :obj:`str` ]: The arguments extracted from the string.
-
-        Raises:
-            :obj:`~.errors.UnclosedQuotes`
+            Tuple[ List[ :obj:`str` ], :obj:`str`]: The arguments extracted from the string, as well as any
+                remainder if ``max_parse`` was supplied.
         """
         finished = False
         args_list = []
-        while not finished:
+        while not finished and len(args_list) < max_parse:
             arg = self._next_str()
             if arg is not None:
                 args_list.append(arg)
             else:
                 finished = True
-        return args_list
+        return args_list, self.text[self.index :].strip()