+++ tests/test_commands.py
@@ -70,43 +70,6 @@ def test_Command_add_check_adds_callable_to_list(dummy_command, dummy_function):
     assert dummy_command._checks == [dummy_function]
 
 
-def test_Command_args_before_asterisk_return(dummy_command):
-    args_num, name = dummy_command.arg_details._args_and_name_before_asterisk()
-    assert args_num == 0
-
-
-def test_Command_args_before_asterisk_raise_error():
-    # Check if _args_before_asterisk raises TypeError after having more than 1 arg after asterisk
-
-    def dummy_function(a, b, *, c, d):
-        pass
-
-    with pytest.raises(TypeError) as error:
-        dummy_cmd = commands.Command(dummy_function, "dummy", True, [], False)
-        dummy_cmd.arg_details._args_and_name_before_asterisk()
-
-    assert error.type is TypeError
-
-
-@pytest.mark.asyncio
-async def test_Commands_convert_args_for_asterisk_arg(dummy_ctx):
-    # Check if function works properly with *arg in command's sign
-    def dummy(a, b, *c):
-        assert a == "witcher"
-        assert b == "was"
-        assert c == ("here",)
-
-    args = ["witcher", "was", "here"]
-    cmd = commands.SignatureInspector(commands.Command(dummy, "dummy", False, [], False))
-
-    before_asterisk, name = cmd._args_and_name_before_asterisk()
-    arg_details = list(cmd.args.values())[1 : len(args) + 1]
-
-    new_args = await commands.Command._convert_args(dummy_ctx, args[:before_asterisk], arg_details)
-    new_args = [*new_args, *args[before_asterisk:]]
-    dummy(*new_args)
-
-
 def test_Group_get_subcommand_returns_None(dummy_group):
     assert dummy_group.get_subcommand("foo") is None